requirements/production.txt

Flask==1.1.2
marshmallow==3.7.1
flask-apispec==0.11.0
Flask-SQLAlchemy==2.2
Flask-Injector==0.12.3
zope.sqlalchemy==1.3
transaction==3.0.1
python-dotenv==0.15.0
flask-migrate==2.5.3
Flask-HTTPAuth==4.2.0
requests==2.25.1